+++ src/SimpleVoteManager.sol
@@ -9,16 +9,16 @@ import {UUPSUpgradeable} from "@openzeppelin/contracts/proxy/utils/UUPSUpgradeab
 contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgradeable, UUPSUpgradeable {
     struct Vote {
         address proposer;
-        string  name;
+        string name;
         uint48 startDate;
         uint48 endDate;
         uint48 resultsDate;
         uint48 totalVoters;
         mapping(string => bool) voters;
         string[] options;
-        mapping(string => bool)     existingOptions;
-        mapping(string => uint256)  votes;
-        mapping(string => string)   nullifiers;
+        mapping(string => bool) existingOptions;
+        mapping(string => uint256) votes;
+        mapping(string => string) nullifiers;
     }
 
     mapping(string => Vote) private votes;
@@ -63,11 +63,7 @@ contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUp
         __Ownable_init(initialOwner);
     }
 
-    function _authorizeUpgrade(address newImplementation)
-        internal
-        override
-        onlyOwner
-    {}
+    function _authorizeUpgrade(address newImplementation) internal override onlyOwner {}
 
     function createVote(
         string calldata id,
@@ -89,21 +85,21 @@ contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUp
         votes[id].resultsDate = resultsDate;
         votes[id].totalVoters = uint48(voters.length);
         votes[id].options = options;
-        for (uint i = 0; i < voters.length; i++) {
+        for (uint256 i = 0; i < voters.length; i++) {
             votes[id].voters[voters[i]] = true;
         }
-        for (uint i = 0; i < options.length; i++) {
+        for (uint256 i = 0; i < options.length; i++) {
             votes[id].existingOptions[options[i]] = true;
         }
         emit VoteCreated(id, name);
     }
 
-    function updateVoteDates(
-        string calldata id,
-        uint48 startDate,
-        uint48 endDate,
-        uint48 resultsDate
-    ) external validVoteDates(startDate, endDate, resultsDate) existingVote(id) onlyProposer(id) {
+    function updateVoteDates(string calldata id, uint48 startDate, uint48 endDate, uint48 resultsDate)
+        external
+        validVoteDates(startDate, endDate, resultsDate)
+        existingVote(id)
+        onlyProposer(id)
+    {
         require(block.timestamp < votes[id].startDate - 24 hours, "Too near to vote start date");
 
         votes[id].startDate = startDate;
@@ -111,7 +107,11 @@ contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUp
         votes[id].resultsDate = resultsDate;
     }
 
-    function castVote(string calldata voteId, string calldata optionId, string calldata nullifier) external nonReentrant existingVote(voteId) {
+    function castVote(string calldata voteId, string calldata optionId, string calldata nullifier)
+        external
+        nonReentrant
+        existingVote(voteId)
+    {
         Vote storage vote = votes[voteId];
         require(vote.voters[nullifier], "Not eligible to vote");
         require(block.timestamp >= vote.startDate && block.timestamp <= vote.endDate, "Voting is not active");
@@ -124,19 +124,36 @@ contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUp
         emit Voted(voteId);
     }
 
-    function getVoteResults(string calldata voteId) external view existingVote(voteId) returns (string[] memory options, uint256[] memory voteCounts) {
+    function getVoteResults(string calldata voteId)
+        external
+        view
+        existingVote(voteId)
+        returns (string[] memory options, uint256[] memory voteCounts)
+    {
         Vote storage vote = votes[voteId];
         require(block.timestamp > vote.resultsDate, "Results are not available yet");
 
         options = vote.options;
-        voteCounts = new uint[](vote.options.length);
+        voteCounts = new uint256[](vote.options.length);
 
-        for (uint i = 0; i < vote.options.length; i++) {
+        for (uint256 i = 0; i < vote.options.length; i++) {
             voteCounts[i] = vote.votes[vote.options[i]];
         }
     }
 
-    function getVoteInfo(string calldata voteId) external view existingVote(voteId) returns (string memory name, uint48 startDate, uint48 endDate, uint48 resultsDate, uint48 totalVoters, string[] memory options) {
+    function getVoteInfo(string calldata voteId)
+        external
+        view
+        existingVote(voteId)
+        returns (
+            string memory name,
+            uint48 startDate,
+            uint48 endDate,
+            uint48 resultsDate,
+            uint48 totalVoters,
+            string[] memory options
+        )
+    {
         Vote storage vote = votes[voteId];
         name = vote.name;
         startDate = vote.startDate;
@@ -146,7 +163,12 @@ contract SimpleVoteManager is Initializable, ReentrancyGuardTransient, OwnableUp
         options = vote.options;
     }
 
-    function getOwnVoteInfo(string calldata voteId, string calldata nullifier) external view existingVote(voteId) returns (bool hasVoted, string memory optionVoted) {
+    function getOwnVoteInfo(string calldata voteId, string calldata nullifier)
+        external
+        view
+        existingVote(voteId)
+        returns (bool hasVoted, string memory optionVoted)
+    {
         Vote storage vote = votes[voteId];
         require(bytes(vote.nullifiers[nullifier]).length > 0, "Not voted yet");
