+++ src/old-versions/BackVoteManagerV1.sol
@@ -14,10 +14,10 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         uint48 resultsDate;
         uint48 totalVoters;
         string[] options;
-        mapping(string => bool)    existingOptions;
-        mapping(string => bool)    voters;
+        mapping(string => bool) existingOptions;
+        mapping(string => bool) voters;
         mapping(string => uint256) votes;
-        mapping(string => string)  nullifiers;
+        mapping(string => string) nullifiers;
     }
 
     address private authorizedCaller;
@@ -51,11 +51,7 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         authorizedCaller = _authorizedCaller;
     }
 
-    function _authorizeUpgrade(address newImplementation)
-        internal
-        override
-        onlyOwner
-    {}
+    function _authorizeUpgrade(address newImplementation) internal override onlyOwner {}
 
     function _validVoteDates(uint48 startDate, uint48 endDate, uint48 resultsDate) internal pure {
         require(startDate < endDate, "Start date must be before end date");
@@ -97,21 +93,21 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         votes[id].resultsDate = resultsDate;
         votes[id].totalVoters = uint48(voters.length);
         votes[id].options = options;
-        for (uint i = 0; i < voters.length; i++) {
+        for (uint256 i = 0; i < voters.length; i++) {
             votes[id].voters[voters[i]] = true;
         }
-        for (uint i = 0; i < options.length; i++) {
+        for (uint256 i = 0; i < options.length; i++) {
             votes[id].existingOptions[options[i]] = true;
         }
         emit VoteCreated(id, name);
     }
 
-    function updateVoteDates(
-        string calldata id,
-        uint48 startDate,
-        uint48 endDate,
-        uint48 resultsDate
-    ) external validVoteDates(startDate, endDate, resultsDate) existingVote(id) onlyAuthorizedCaller {
+    function updateVoteDates(string calldata id, uint48 startDate, uint48 endDate, uint48 resultsDate)
+        external
+        validVoteDates(startDate, endDate, resultsDate)
+        existingVote(id)
+        onlyAuthorizedCaller
+    {
         require(block.timestamp < votes[id].startDate - 24 hours, "Too near to vote start date");
 
         votes[id].startDate = startDate;
@@ -119,7 +115,12 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         votes[id].resultsDate = resultsDate;
     }
 
-    function castVote(string calldata voteId, string calldata optionId, string calldata nullifier) external nonReentrant existingVote(voteId) onlyAuthorizedCaller {
+    function castVote(string calldata voteId, string calldata optionId, string calldata nullifier)
+        external
+        nonReentrant
+        existingVote(voteId)
+        onlyAuthorizedCaller
+    {
         Vote storage vote = votes[voteId];
         require(vote.voters[nullifier], "Not eligible to vote");
         require(block.timestamp >= vote.startDate && block.timestamp <= vote.endDate, "Voting is not active");
@@ -132,19 +133,36 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         emit Voted(voteId);
     }
 
-    function getVoteResults(string calldata voteId) external view existingVote(voteId) returns (string[] memory options, uint256[] memory voteCounts) {
+    function getVoteResults(string calldata voteId)
+        external
+        view
+        existingVote(voteId)
+        returns (string[] memory options, uint256[] memory voteCounts)
+    {
         Vote storage vote = votes[voteId];
         require(block.timestamp > vote.resultsDate, "Results are not available yet");
 
         options = vote.options;
-        voteCounts = new uint[](vote.options.length);
+        voteCounts = new uint256[](vote.options.length);
 
-        for (uint i = 0; i < vote.options.length; i++) {
+        for (uint256 i = 0; i < vote.options.length; i++) {
             voteCounts[i] = vote.votes[vote.options[i]];
         }
     }
 
-    function getVoteInfo(string calldata voteId) external view existingVote(voteId) returns (string memory name, uint48 startDate, uint48 endDate, uint48 resultsDate, uint48 totalVoters, string[] memory options) {
+    function getVoteInfo(string calldata voteId)
+        external
+        view
+        existingVote(voteId)
+        returns (
+            string memory name,
+            uint48 startDate,
+            uint48 endDate,
+            uint48 resultsDate,
+            uint48 totalVoters,
+            string[] memory options
+        )
+    {
         Vote storage vote = votes[voteId];
         name = vote.name;
         startDate = vote.startDate;
@@ -154,7 +172,12 @@ contract BackVoteManager is Initializable, ReentrancyGuardTransient, OwnableUpgr
         options = vote.options;
     }
 
-    function getOwnVoteInfo(string calldata voteId, string calldata nullifier) external view existingVote(voteId) returns (bool hasVoted, string memory optionVoted) {
+    function getOwnVoteInfo(string calldata voteId, string calldata nullifier)
+        external
+        view
+        existingVote(voteId)
+        returns (bool hasVoted, string memory optionVoted)
+    {
         Vote storage vote = votes[voteId];
         require(bytes(vote.nullifiers[nullifier]).length > 0, "Not voted yet");
